server/requirements.txt
flask
python-dotenv
flask-smorest
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
passlib
gunicorn
psycopg2